Turns out the discount on an invoice line item may be a decimal number.  Who knew?
git-svn-id: http://freshbooks-api-java-client.googlecode.com/svn/trunk@25 c2e10e96-c8c8-11dd-a412-a1792178251d

+++ freshbooksApiClient/src/com/freshbooks/model/Invoice.java
@@ -25,7 +25,7 @@ public class Invoice implements Serializable {
     @XStreamAlias("po_number")
     String poNumber;
     
-    Integer discount;
+    Double discount;
     String notes;
     String terms;
     
@@ -97,10 +97,10 @@ public class Invoice implements Serializable {
     public void setPoNumber(String poNumber) {
         this.poNumber = poNumber;
     }
-    public Integer getDiscount() {
+    public Double getDiscount() {
         return discount;
     }
-    public void setDiscount(Integer discount) {
+    public void setDiscount(Double discount) {
         this.discount = discount;
     }
     public String getNotes() {
